Use `DefaultImageLoadingHandler` to load Merino suggestion icons.
The new favicon fetcher doesn't handle the case where it's given an icon
URL, and tries to compose one using the first character of the icon URL
instead. For prototyping, we can load the icon and set it on the view
manually, as Nish suggested in
https://mozilla.slack.com/archives/C4DK7829K/p1674621762334769?thread_ts=1674607956.502259&cid=C4DK7829K.

Loading the icon happens asynchronously, so we must also take care not
to set the loaded icon on the view if the cell has already been reused.
Otherwise, we'll end up setting the wrong image on the reused cell,
which is a common bug in UIKit programming:
https://medium.com/ios-seminar/why-we-use-dequeuereusablecellwithidentifier-ce7fd97cde8e#562e